ci: #1333 add update release note pipeline
